requirements.txt
streamlit
opencv-python
codetiming
pillow
dalle-mini
git+https://github.com/patil-suraj/vqgan-jax.git